Feat: Binding Auth0 to local accounts
Implemented user context function that can be run in
authorized endpoints. This will determine if the logged
in auth0 user has a local account. If it does not it will
create a local account and generate a unique username.
The function then returns a dict containing user-info,
which can be used to validate certain requests.